Only create a view in -loadView if no nib-based view is present
The default UIViewController#loadView method will load a a view from a
nib file if one is present and connected to the view outlet.  If a nib
is present, the 'nibName' propery will be set to a non-nil value.

According to Apple, if you're loading your view from a nib, you should
not override loadView, but mimicing the default -loadView behavior by
exiting if a nib-based view is present seems to be fine.